StackExchange/requirements.txt
Package                   Version   
------------------------- ----------  
Django                    2.1.1     
django-cors-headers       3.1.0     
django-rest-passwordreset 1.0.0     
djangorestframework       3.10.3    
djangorestframework-jwt   1.11.0     
future                    0.16.0        
pip                       19.3.1    
pycryptodome              3.9.0     
PyJWT                     1.7.1     
pylint                    2.4.4     
pypng                     0.0.20    
python-dateutil           2.8.0     
pytz                      2019.3    
PyYAML                    3.13      
reportlab                 3.5.34    
requests                  2.20.1       
setuptools                42.0.2    
six                       1.13.0        
urllib3                   1.24.3        
wheel                     0.33.6    
wrapt                     1.11.2    
